+++ package.json
@@ -17,7 +17,9 @@
         "ts-node": "^10.9.1",
         "typescript": "^5.2.2",
         "vite": "^4.4.9",
-        "web-ext": "^7.7.0"
+        "web-ext": "^7.7.0",
+        "vue-tsc": "^1.8.8",
+        "@vitejs/plugin-vue": "^4.2.3"
     },
     "scripts": {
         "build": "yarn build:firefox",
@@ -43,6 +45,7 @@
         "@types/chrome": "^0.0.245",
         "@types/webextension-polyfill": "^0.10.2",
         "bootstrap-icons": "^1.10.5",
-        "webextension-polyfill": "^0.10.0"
+        "webextension-polyfill": "^0.10.0",
+        "vue": "^3.3.4"
     }
 }
