+++ src/shared/options/components/ThreeColorSetting.vue
@@ -12,8 +12,6 @@
 <script lang="ts">
 import { defineComponent } from "vue";
 
-import { zBtnColors, zIcnColors } from "../../data/type";
-
 import { TUICData } from "../../../content/data";
 import { TUICLibrary } from "../../../content/library";
 import { TUICPref } from "../../../content/library";
@@ -71,9 +69,8 @@ import ColorSetting from "./ColorSetting.vue";
 
 export default defineComponent({
     setup(props) {
-        const isBtnColors = zBtnColors.safeParse(TUICData.colors[props.id]).success;
-        const isIcnColors = zIcnColors.safeParse(TUICData.colors[props.id]).success;
-        console.log(isBtnColors);
+        const isBtnColors = !!TUICData.colors[props.id]["background"];
+        const isIcnColors = !TUICData.colors[props.id]["background"];
         return { isBtnColors, isIcnColors, TUICLibrary, TUICPref };
     },
     props: {
